KAS-5052 Review of changes, route change needed more updates
