+++ app/components/cases/submissions-header.hbs
@@ -10,7 +10,7 @@
       </Group.Item>
       <Group.Item>
         <Auk::Tabs @responsive={{true}} @responsiveBreakpoint="large">
-          <Auk::Tab @route="submissions.index">
+          <Auk::Tab @route="submissions.ongoing">
             {{t "to-follow-up"}}
           </Auk::Tab>
           {{#if (user-may "view-concept-submissions")}}
