+++ app/routes/cases/case/subcases/subcase/new-submission.js
@@ -26,7 +26,7 @@ export default class CasesCaseSubcasesSubcaseNewSubmissionRoute extends Route {
   async beforeModel(_transition) {
     if (!this.currentSession.may('create-submissions')) {
       if (this.currentSession.may('view-submissions')) {
-        return this.router.transitionTo('submissions');
+        return this.router.transitionTo('submissions.ongoing');
       }
       return this.router.transitionTo('cases.index');
     }
