+++ app/routes/submissions.js
@@ -6,9 +6,6 @@ export default class SubmissionsRoute extends Route {
   @service router;
   @service('session') simpleAuthSession;
 
-  // We only want to load the defaults once. Changes made after stay untill all checkboxes are cleared.
-  loadedDefaults = false;
-
   async beforeModel(transition) {
     const isAuthenticated = this.simpleAuthSession.requireAuthentication(transition, this.simpleAuthSession.unauthenticatedRouteName);
