+++ app/templates/submissions.hbs
@@ -1,2 +1,3 @@
 {{page-title (t "submissions")}}
+<Cases::SubmissionsHeader/>
 {{outlet}}
